Added automatic compilation to standard export scripts

+++ example_projects/export_test.py
@@ -2,10 +2,23 @@
 import sys
 sys.path.append('..')
 import main
+import optparse
+import os
 import shutil
 
-export_dir = 'default_src'
-xml_file = 'ruo2.xml'
+
+usage = 'usage: %prog [xml file] [export dir]\n'\
+    + '    export directory will be deleted first!'
+
+parser = optparse.OptionParser(usage=usage)
+(option, args) = parser.parse_args()
+
+if len(args) != 2 :
+    parser.print_help()
+    exit()
+
+xml_file = args[0]
+export_dir = args[1]
 
 shutil.rmtree(export_dir,ignore_errors=True)
 kmc = main.KMC_Editor()
@@ -13,4 +26,7 @@ kmc.project_tree.import_xml_file(xml_file)
 
 kmc.on_btn_export_src__clicked('',export_dir)
 
+if os.uname()[0] == 'Linux':
+    os.chdir(export_dir)
+    os.system("./compile_for_f2py" )
 exit()
